2048 Attempt 1
https://www.codingame.com/ide/puzzle/2048

Score was 62,000 or something. 
I placed 630.

I will optimize this code.
Currently it's just seeing if it can combine two tiles. If it can, it does.
If it cannot, it tries to see where it can move with priority to up and left.

Much improvement can be made.

Detecting new tiles, checking for highest tile and subsequently high tiles, keeping them in the upper left with highest priority